# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_gbaski_gbaski-host_internal_common.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      otp_required:
        type: boolean
      status:
        example: success
        type: string
    type: object
...
  payout.ChangePayoutTypeRequest:
    properties:
      payoutAccountId:
        type: integer
      payoutType:
        type: string
      userId:
        type: string
    type: object
  payout.CreatePaymentPayoutRequest:
    properties:
      provider:
        type: string
      publicKey:
        type: string
      secretKey:
        type: string
      sellerName:
        type: string
      state:
        type: string
      userId:
        type: string
    type: object
  payout.CreatePayoutAccountRequest:
    properties:
      accountName:
        type: string
      accountNumber:
        type: string
      bankId:
        type: integer
      currency:
        type: string
      userId:
        type: string
    type: object
  payout.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      otp_required:
        type: boolean
      status:
        example: success
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@gbaski.app
    name: API Support
  description: API documentation for Gbaski Host service
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Gbaski Host API
  version: "1.0"
paths:
...
  /payout/account:
    get:
      consumes:
      - application/json
      description: Retrieve the payout account for the authenticated user
      parameters:
      - default: NGN
        description: Currency code
        in: query
        name: currency
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_common.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_common.Response'
      security:
      - Bearer: []
      summary: Get payout account
      tags:
      - payout
  /payout/banks:
    get:
      consumes:
      - application/json
      description: Retrieve a list of available banks for payout account setup
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payout.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/payout.Response'
      security:
      - Bearer: []
      summary: Get payout banks
      tags:
      - payout
  /payout/change-type:
    post:
      consumes:
      - application/json
      description: Change the payout type for a payout account
      parameters:
      - description: Payout type change request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/payout.ChangePayoutTypeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payout.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/payout.Response'
      security:
      - Bearer: []
      summary: Change payout type
      tags:
      - payout
  /payout/create-account:
    post:
      consumes:
      - application/json
      description: Create a new payout account for the authenticated user with OTP
        verification
      parameters:
      - description: OTP code for verification
        in: query
        name: otp_code
        type: string
      - description: Payout account creation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/payout.CreatePayoutAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payout.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/payout.Response'
      security:
      - Bearer: []
      summary: Create payout account
      tags:
      - payout
  /payout/history/{formId}:
    get:
      consumes:
      - application/json
      description: Retrieve paginated payout history for a specific form
      parameters:
      - description: Form ID (UUID)
        in: path
        name: formId
        required: true
        type: string
      - description: Search query
        in: query
        name: search
        type: string
      - description: Filter parameters
        in: query
        name: filter
        type: string
      - description: Range parameters
        in: query
        name: range
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payout.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/payout.Response'
      security:
      - Bearer: []
      summary: Get payout history
      tags:
      - payout
  /payout/payment-provider:
    get:
      consumes:
      - application/json
      description: Retrieve payment provider integration details for the authenticated
        user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payout.Response'
      security:
      - Bearer: []
      summary: Get payment payout provider
      tags:
      - payout
    post:
      consumes:
      - application/json
      description: Create or update payment provider integration for payouts
      parameters:
      - description: Payment payout provider details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/payout.CreatePaymentPayoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payout.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/payout.Response'
      security:
      - Bearer: []
      summary: Create payment payout provider
      tags:
      - payout
  /payout/summary/{formId}:
    get:
      consumes:
      - application/json
      description: Retrieve payout summary for a specific form
      parameters:
      - description: Form ID (UUID)
        in: path
        name: formId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payout.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/payout.Response'
      security:
      - Bearer: []
      summary: Get payout summary
      tags:
      - payout
  /payout/total-payout/{formId}:
    get:
      consumes:
      - application/json
      description: Retrieve total payout amount for a specific form
      parameters:
      - description: Form ID (UUID)
        in: path
        name: formId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payout.Response'
      security:
      - Bearer: []
      summary: Get total payout
      tags:
      - payout
...
schemes:
- http
- https
swagger: "2.0"